% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSAT_tidy}
\alias{LSAT_tidy}
\title{Tidy version of \link{ltm}'s LSAT dataset}
\format{A data frame with columns:
\describe{
 \item{testid}{Unique ID for each test, there is only one test in this dataset}
 \item{userid}{Unique ID for each user}
 \item{itemid}{Unique ID for each item}
 \item{response}{User's response to item, 0 or 1}
}}
\source{
R package \link{ltm}
}
\usage{
LSAT_tidy
}
\description{
Tidy version of \link{ltm}'s LSAT dataset
}
\examples{
\dontrun{
 LSAT_tidy
}
}
\keyword{datasets}
