% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prop_value}
\alias{prop_value}
\title{Find proportion of a value in vector}
\usage{
prop_value(x, val, na.rm = TRUE)
}
\arguments{
\item{x}{An atomic vector.}

\item{val}{A value belonging to vector.}

\item{na.rm}{Remove \code{NA} values?}
}
\value{
proportion of \code{val} in \code{x}
}
\description{
Find proportion of a value in vector
}
